Remove old logic
