+++ pages/index.tsx
@@ -1,8 +1,20 @@
-import { convertTypesAndGetEdges } from "helpers/getEdgeNodesHelpers";
+import getFirstParagraph from "helpers/getFirstParagraph";
 import { PostDocumentWithoutBody } from "interfaces";
-import type { GetStaticProps, NextPage } from "next";
+import type { NextPage } from "next";
+import Head from "next/head";
+import { setHostUrl } from "redux/slices/hostUrl";
+import { setTinaData } from "redux/slices/tinaData";
+import { useAppDispatch, useAppSelector } from "redux/typesHooks";
+import { useTina } from "tinacms/dist/react";
 
-import HomePage from "@/components/HomePage";
+import Footer from "@/components/Footer";
+import Header from "@/components/Header";
+import Intro from "@/components/Intro";
+import Pagination from "@/components/Pagination";
+import Posts from "@/components/Posts";
+import StandWithUkraine from "@/components/StandWithUkraine";
+import Tags, { Tag } from "@/components/Tags";
+import useIsomorphicLayoutEffect from "@/components/useIsomorphicLayoutEffect";
 
 import client from ".tina/__generated__/client";
 import {
@@ -10,38 +22,122 @@ import {
   MainConfigQueryVariables,
 } from ".tina/__generated__/types";
 
-const Home: NextPage<{
+interface Props {
   posts: PostDocumentWithoutBody[];
   tinaData: MainConfigQuery;
   query: string;
   variables: MainConfigQueryVariables;
-}> = ({ posts, tinaData, query, variables }) => (
-  <HomePage
-    homePagePosts={posts}
-    tinaData={tinaData}
-    query={query}
-    variables={variables}
-  />
-);
+}
 
-export const getStaticProps: GetStaticProps = async () => {
-  const posts = await client.queries.postsWithoutBody({
-    filter: {
-      draft: { eq: false },
-      date: { before: new Date(Date.now()).toString() },
-    },
+const initialTagCount: Record<string, number> = {};
+
+const calculateSortedTags = (posts: PostDocumentWithoutBody[]): Tag[] => {
+  const tags = posts.map((post) => post.tags).flat();
+
+  const tagsFrequency = tags.reduce((acc, tag) => {
+    if (acc[tag]) {
+      return { ...acc, [tag]: acc[tag] + 1 };
+    }
+
+    return { ...acc, [tag]: 1 };
+  }, initialTagCount);
+
+  return Object.entries(tagsFrequency)
+    .map(([name, postsCount]) => ({
+      name,
+      postsCount,
+    }))
+    .sort((a, b) => b.postsCount - a.postsCount);
+};
+
+const Home: NextPage<Props> = ({ posts, query, tinaData, variables }) => {
+  const dispatch = useAppDispatch();
+
+  const { data } = useTina({
+    query,
+    variables,
+    data: tinaData,
   });
 
-  const { data, query, variables } = await client.queries.mainConfig({
+  const siteDescription = getFirstParagraph(data.mainConfig.siteDescription);
+
+  const selectedTags = useAppSelector((state) => state.selectedTags);
+  const currentPage = useAppSelector((state) => state.pagination.currentPage);
+  const nextPage = currentPage + 1;
+
+  const tags = calculateSortedTags(posts);
+
+  const filteredPosts = posts.filter((post) => {
+    if (selectedTags.length === 0) {
+      return true;
+    }
+
+    return selectedTags.some((tag) => post.tags.includes(tag));
+  });
+
+  const pagesCount = Math.ceil(
+    filteredPosts.length / data.mainConfig.postsPerPage
+  );
+
+  const postsToRender = filteredPosts.slice(
+    currentPage * data.mainConfig.postsPerPage,
+    nextPage * data.mainConfig.postsPerPage
+  );
+
+  useIsomorphicLayoutEffect(() => {
+    dispatch(setHostUrl(window.location.origin));
+    dispatch(setTinaData(data));
+  }, []);
+
+  const hostURLLink = useAppSelector((state) => state.hostUrl);
+
+  return (
+    <>
+      <Head>
+        <title>{data.mainConfig.siteTitle}</title>
+        <meta property="og:title" content={data.mainConfig.ogSiteTitle} />
+        <meta property="og:description" content={siteDescription} />
+        <meta property="og:url" content={hostURLLink} />
+        <meta property="og:type" content="website" />
+        <meta property="og:site_name" content={data.mainConfig.siteTitle} />
+        <meta property="og:image" content={data.mainConfig.defaultImage} />
+        <meta name="description" content={siteDescription} />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+      <StandWithUkraine />
+      <Header />
+      <Intro
+        authorName={data.mainConfig.authorName}
+        authorPosition={data.mainConfig.authorPosition}
+        siteDescription={data.mainConfig.siteDescription}
+      />
+      <section className="simple-section">
+        <div className="container">
+          <Tags tags={tags} />
+          <Posts posts={postsToRender} />
+          <Pagination pagesCount={pagesCount} />
+        </div>
+      </section>
+      <Footer />
+    </>
+  );
+};
+
+export const getStaticProps = async () => {
+  const posts = await client.queries.postConnection({});
+
+  const mainConfig = await client.queries.mainConfig({
     relativePath: "mainConfig.json",
   });
 
   return {
     props: {
-      posts: convertTypesAndGetEdges(posts),
-      tinaData: data,
-      query,
-      variables,
+      posts: posts.data.postConnection.edges
+        ?.map((edge) => edge?.node)
+        .reverse(),
+      tinaData: mainConfig.data,
+      query: mainConfig.query,
+      variables: mainConfig.variables,
     },
   };
 };
